Тесты для ДЗ 3 недели
